src/Savin_P3369_Lab1(old)/requirements.txt
pillow>=8.0.0
matplotlib>=3.3.0
numpy>=1.19.0